Include precision and unit of measurement on entity state
