+++ Sources/Extensions/AppIntents/Widget/Sensor/WidgetSensorsAppIntentTimelineProvider.swift
@@ -87,94 +87,21 @@ struct WidgetSensorsAppIntentTimelineProvider: AppIntentTimelineProvider {
         for sensor: IntentSensorsAppEntity,
         server: Server
     ) async throws -> WidgetSensorsEntry.SensorData {
-        guard let connection = Current.api(for: server)?.connection else {
-            Current.Log.error("No API available to fetch sensor data")
-            throw HomeAssistantAPI.APIError.noAPIAvailable
-        }
-
-        let result = await withCheckedContinuation { continuation in
-            connection.send(.init(
-                type: .rest(.get, "states/\(sensor.entityId)"),
-                shouldRetry: true
-            )) { result in
-                continuation.resume(returning: result)
-            }
-        }
-
-        var data: HAData?
-        switch result {
-        case let .success(resultData):
-            data = resultData
-        case let .failure(error):
-            Current.Log.error("Failed to render template for details widget: \(error)")
-            throw WidgetSensorsDataError.apiError
-        }
-
-        guard let data else {
-            throw WidgetSensorsDataError.apiError
-        }
-
-        var state: [String: Any]?
-        switch data {
-        case let .dictionary(response):
-            state = response
-        default:
-            Current.Log.error("Failed to render template for sensor widget: Bad response data")
-            throw WidgetSensorsDataError.badResponse
-        }
-
-        var stateValue = (state?["state"] as? String) ?? "N/A"
-        stateValue = adjustPrecision(
-            serverId: server.identifier.rawValue,
-            entityId: sensor.entityId,
-            stateValue: stateValue
-        )
-        let unitOfMeasurement = (state?["attributes"] as? [String: Any])?["unit_of_measurement"] as? String
-        stateValue = capitalizeFirstLetter(stateValue)
+        let state: ControlEntityProvider.State = await ControlEntityProvider(domains: Domain.allCases).state(
+            server: server,
+            entityId: sensor.entityId
+        ) ?? ControlEntityProvider.State(value: "", unitOfMeasurement: nil)
         return WidgetSensorsEntry.SensorData(
             id: sensor.id,
             key: sensor.displayString,
-            value: stateValue,
-            unitOfMeasurement: unitOfMeasurement,
+            value: state.value,
+            unitOfMeasurement: state.unitOfMeasurement,
             icon: sensor.icon
         )
     }
 
     private func capitalizeFirstLetter(_ text: String) -> String {
         text.prefix(1).uppercased() + text.dropFirst()
-    }
-
-    private func adjustPrecision(serverId: String, entityId: String, stateValue: String) -> String {
-        guard let stateValueFloat = Float(stateValue) else {
-            return stateValue
-        }
-        if let decimalPlacesForEntityId: Int = {
-            do {
-                return try Current.database.read { db in
-                    try AppEntityRegistryListForDisplay
-                        .filter(
-                            Column(DatabaseTables.AppEntityRegistryListForDisplay.id.rawValue) == ServerEntity
-                                .uniqueId(
-                                    serverId: serverId,
-                                    entityId: entityId
-                                )
-                        )
-                        .fetchOne(db)?.registry.decimalPlaces
-                }
-            } catch {
-                Current.Log.error("Failed to fetch decimal places for entity ID: \(entityId)")
-                return nil
-            }
-        }() {
-            let numberFormatter = NumberFormatter()
-            numberFormatter.numberStyle = .decimal
-            numberFormatter.locale = Locale.current
-            numberFormatter.maximumFractionDigits = decimalPlacesForEntityId
-            numberFormatter.minimumFractionDigits = decimalPlacesForEntityId
-            return numberFormatter.string(from: NSNumber(value: stateValueFloat)) ?? stateValue
-        } else {
-            return stateValue
-        }
     }
 
     private func suggestions() async -> [(Server, [HAAppEntity])] {
