+++ Sources/Shared/ControlEntityProvider.swift
@@ -1,5 +1,6 @@
 import Foundation
 import GRDB
+import HAKit
 
 public final class ControlEntityProvider {
     public enum States: String {
@@ -9,6 +10,16 @@ public final class ControlEntityProvider {
         case closing
         case on
         case off
+    }
+
+    public struct State {
+        public let value: String
+        public let unitOfMeasurement: String?
+
+        public init(value: String, unitOfMeasurement: String?) {
+            self.value = value
+            self.unitOfMeasurement = unitOfMeasurement
+        }
     }
 
     public let domains: [Domain]
@@ -62,5 +73,53 @@ public final class ControlEntityProvider {
         }
 
         return entitiesPerServer
+    }
+
+    public func state(server: Server, entityId: String) async -> State? {
+        guard let connection = Current.api(for: server)?.connection else {
+            Current.Log.error("No API available to fetch state data")
+            return nil
+        }
+
+        let result = await withCheckedContinuation { continuation in
+            connection.send(.init(
+                type: .rest(.get, "states/\(entityId)"),
+                shouldRetry: true
+            )) { result in
+                continuation.resume(returning: result)
+            }
+        }
+
+        var data: HAData?
+        switch result {
+        case let .success(resultData):
+            data = resultData
+        case let .failure(error):
+            Current.Log.error("Failed to get state: \(error)")
+            return nil
+        }
+
+        guard let data else {
+            return nil
+        }
+
+        var state: [String: Any]?
+        switch data {
+        case let .dictionary(response):
+            state = response
+        default:
+            Current.Log.error("Failed to get state bad response data")
+            return nil
+        }
+
+        var stateValue = (state?["state"] as? String) ?? "N/A"
+        stateValue = StatePrecision.adjustPrecision(
+            serverId: server.identifier.rawValue,
+            entityId: entityId,
+            stateValue: stateValue
+        )
+        let unitOfMeasurement = (state?["attributes"] as? [String: Any])?["unit_of_measurement"] as? String
+        stateValue = stateValue.capitalizedFirst
+        return .init(value: stateValue, unitOfMeasurement: unitOfMeasurement)
     }
 }
